API

Numerical methods

ConstrainedRootSolvers.find_zero — Function

Function to find the first root that gives a target function result of zero. If the root does not exist, the function returns the point where the target function is most close to zero.

find_zero(f, ms, tol; stepping)

defined at /home/runner/work/ConstrainedRootSolvers.jl/ConstrainedRootSolvers.jl/src/find_zero.jl:40.

find_zero(f, ms, tol; stepping)

defined at /home/runner/work/ConstrainedRootSolvers.jl/ConstrainedRootSolvers.jl/src/find_zero.jl:144.

find_zero(f, ms, tol; stepping)

defined at /home/runner/work/ConstrainedRootSolvers.jl/ConstrainedRootSolvers.jl/src/find_zero.jl:258.

find_zero(f, ms, tol; stepping)

defined at /home/runner/work/ConstrainedRootSolvers.jl/ConstrainedRootSolvers.jl/src/find_zero.jl:322.

ConstrainedRootSolvers.find_peak — Function

find_peak(f::F,
          ms::AbstractCRSMethod{FT},
          tol::AbstractTolerance{FT}
) where {F<:Function, FT<:AbstractFloat}

Find the solution, given

f A function to solve

ms BisectionMethod type method struct

tol AbstractTolerance type tolerance struct

Note that to compute the lowest value, use -f to make it a peak.

Possible combinations

Method Type	Tolerance type	Pass Test

BisectionMethod	SolutionTolerance	2 allocs

NelderMeadMethod	ResidualTolerance	N+6 allocs

NelderMeadMethod	SolutionToleranceND	N+6 allocs

ReduceStepMethod	SolutionTolerance	Yes

ReduceStepMethodND	SolutionToleranceND	4 allocs

Method options

ConstrainedRootSolvers.AbstractCRSMethod — Type

Abstract type of the ConstrainedRootSolvers methods

abstract type AbstractCRSMethod{FT<:AbstractFloat}

ConstrainedRootSolvers.BisectionMethod — Type

Bisection method for 1D root solvers

mutable struct BisectionMethod{FT<:AbstractFloat} <: ConstrainedRootSolvers.AbstractCRSMethod{FT<:AbstractFloat}

Fields

x_min::AbstractFloat

lower bound

x_max::AbstractFloat

upper bound

xy::Matrix{FT} where FT<:AbstractFloat

matrix that stores x and y

history::Vector{T} where T

history of all simulations

ConstrainedRootSolvers.NelderMeadMethod — Type

Nelder-Mead method for 2D and above solvers

mutable struct NelderMeadMethod{FT<:AbstractFloat} <: ConstrainedRootSolvers.AbstractCRSMethod{FT<:AbstractFloat}

Fields

N::Int64

Number of parameters to optimize

x_inis::Vector{FT} where FT<:AbstractFloat

Initial values

simplex::Array{Vector{FT}, 1} where FT<:AbstractFloat

Simplex array of array with dimension (N+1) * (N+1)

cen_x::Vector{FT} where FT<:AbstractFloat

Centroid

ref_x::Vector{FT} where FT<:AbstractFloat

Reflection

exp_x::Vector{FT} where FT<:AbstractFloat

Expansion

con_x::Vector{FT} where FT<:AbstractFloat

Contraction

ConstrainedRootSolvers.NewtonBisectionMethod — Type

Newton's method constrained by mininum and maximum ranges for 1D root solver

mutable struct NewtonBisectionMethod{FT<:AbstractFloat} <: ConstrainedRootSolvers.AbstractCRSMethod{FT<:AbstractFloat}

Fields

x_min::AbstractFloat

Lower bound

x_max::AbstractFloat

Upper bound

x_ini::AbstractFloat

Initial guess

history::Vector{T} where T

history of all simulations

ConstrainedRootSolvers.NewtonRaphsonMethod — Type

Newton raphson method for 1D root solver

mutable struct NewtonRaphsonMethod{FT<:AbstractFloat} <: ConstrainedRootSolvers.AbstractCRSMethod{FT<:AbstractFloat}

Fields

x_ini::AbstractFloat

Initial guess

history::Vector{T} where T

history of all simulations

ConstrainedRootSolvers.ReduceStepMethod — Type

Reduce step method for 1D root solver. This method increases or decreases from initial guess until no improvement is found. Then the incremantal step decreases, and then the root solver continues.

mutable struct ReduceStepMethod{FT<:AbstractFloat} <: ConstrainedRootSolvers.AbstractCRSMethod{FT<:AbstractFloat}

Fields

x_min::AbstractFloat

Lower bound

x_max::AbstractFloat

Upper bound

x_ini::AbstractFloat

Initial guess

Δ_ini::AbstractFloat

Initial step

history::Vector{T} where T

history of all simulations

ConstrainedRootSolvers.ReduceStepMethodND — Type

Reduce step method for 2D and above root solver. This method increases or decreases each variable in the initial guess until no improvement is found. Then the incremental steps decreases, and then the root solver continues.

mutable struct ReduceStepMethodND{FT<:AbstractFloat} <: ConstrainedRootSolvers.AbstractCRSMethod{FT<:AbstractFloat}

Fields

x_mins::Vector{FT} where FT<:AbstractFloat

Lower bound

x_maxs::Vector{FT} where FT<:AbstractFloat

Upper bound

x_inis::Vector{FT} where FT<:AbstractFloat

Initial guess

x_targ::Vector{FT} where FT<:AbstractFloat

Target x

x_temp::Vector{FT} where FT<:AbstractFloat

Temporary x

Δ_inis::Vector{FT} where FT<:AbstractFloat

Initial step

Δ_oper::Vector{FT} where FT<:AbstractFloat

Operation step

Δjd::Vector{Bool}

Array of judges

Tolerance options

ConstrainedRootSolvers.AbstractTolerance — Type

Abstract tolerance type

abstract type AbstractTolerance{FT}

ConstrainedRootSolvers.ResidualTolerance — Type

Tolerance for target function residual

struct ResidualTolerance{FT} <: ConstrainedRootSolvers.AbstractTolerance{FT}

Fields

tol::Any

Tolerance for residual

n_limit::Int64

limit of iterations

ConstrainedRootSolvers.SolutionTolerance — Type

Tolerance for solution

struct SolutionTolerance{FT} <: ConstrainedRootSolvers.AbstractTolerance{FT}

Fields

tol::Any

Tolerance for solution

n_limit::Int64

limit of iterations

ConstrainedRootSolvers.SolutionToleranceND — Type

Tolerance for 2D and above solution

struct SolutionToleranceND{FT} <: ConstrainedRootSolvers.AbstractTolerance{FT}

Fields

tol::Vector{FT} where FT

Tolerance for solution

n_limit::Int64

limit of iterations

ConstrainedRootSolvers.if_break — Function

Determine whether to stopping finding the solution depending on the tolerance type.

if_break(tol, x1, x2, y, n)

defined at /home/runner/work/ConstrainedRootSolvers.jl/ConstrainedRootSolvers.jl/src/tolerance.jl:118.

if_break(tol, x1, x2, y, n)

defined at /home/runner/work/ConstrainedRootSolvers.jl/ConstrainedRootSolvers.jl/src/tolerance.jl:148.

ConstrainedRootSolvers.next_xy! — Function

next_xy!(f::F, xy::Array{FT,2}) where {F<:Function, FT<:AbstractFloat}

Determine the next points to, given

f Function to find peak

xy Array of x and y